Ensure apt-cache updates in container cache if debian/control changed
Every time `apt-get install` runs there is a risk that the apt-cache was
outdated and downloads fail on 404 errors. Hence, ensure that all
Containerfile lines that run apt start first with refreshing the cache.

Also make the apt-get invocations slightly less verbose with one `-q` and
DEBUG runs more verbose to show if container build was run with extra args.