Use 'gbp --git-ignore-branch' for better compatibility
Projects that are in git but do not use git-buildpackage might have branch
name 'main' or 'master' or whatever. Requiring builds to run off 'master'
or 'debian/latest' (whatever is gbp default or configured in gbp.conf)
does not make sense. If a user runs 'debcraft build', assume they want to
specifically build the branch that is open and thus ignore branch names.