Ensure diffoscope never runs for more than 5 minutes
Diffoscope is great in detecting and showing potentially unexpected small
changes between two builds that have very few code changes. However, if
the changes are extensive, diffoscope is pretty useless. To strike a
balance of some sorts for now, abort diffoscope after 5 minutes, or force
it to stop after 8 minutes. This way it will not run unreasonably long
trying to compare too large of differences as the report would be hard
or impossible to read anyway.