Replace Ctrl+C bailout in PPA upload with a clear yes/no prompt
Using Ctrl+C resulted in the whole script bailing out. Wit this yes/no
prompt one can still proceed with upload by simply pressing enter
(no typing of 'y' is even required as it is the default) but it one selects
'no' it will continue to execute the rest of the script.