Compare to previous tagged release or build
Current build artifact and log comparisons only compare to the previous
build, which mainly helps discover what effects each commit/change has.

However, if one wants to review what are the changes of the current commit
in comparison to the previously released version, one should be comparing
to the last tagged build/release.

This change has the git tag detection correctly, but comparison code is
just duplicating previous release logic and should be refactored.

Also the comparison of build artifacts would also further needed to be
normalized on version strings, as otherwise the diff does not know that
it indeed is comparing the different versions of the same files. This is
however good enough for now, and ensures all new builds have some reference
to compare against, as before this change the first build on a new branch
name had no comparisons at all.