Omit run-time DEB_BUILD_OPTIONS contents from man page (Closes: 1077479)
The man page is supposed to be static and same for everyone. Thus the
showing the value of DEB_BUILD_OPTIONS at run-time should be omitted from
the man page, and only shows for the specific individual `debcraft --help`
invocations.

Having variable contents in the man page also prevent reproducible builds,
and indeed this issue was noticed there.