Resolve Debcraft version at run-time (Closes: #1077627)
Instead of a static version string in the script, resolve the version at
run-time from the Debian package changelog, or if running from git, from
git tag and commit id.

This obsoletes the need to maintain a static version string in the script
itself and ensures script will never report an outdated versio in the
published Debian package.