Refactor Containerfile to be compatible with older Debian releases
Until Debputy and dependencies are available in older Debian and Ubuntu
releases, the installation needs to be a separate step that is allowed
to fail. Functionality that depends on Debputy will obviously not work,
but at least `debcraft build` and `debcraft release` will be usable on
packages targeting older releases.