Always run `apt-get update` before installing packages in containers
If the apt-cache is not recent, the `apt-get install` commands will fail
on 404 errors. Updating the cache does not take long if a previous layer
already pulled it, so best to always run it just to be sure.

Also indent RUN commands neatly with the apt-get parameters indented one
more level to show are part of the same command.