Apply patches with Quilt when dpkg-source is skipped (Closes: #11, #13)
Normally dpkg-source applies the patches when creating a source package.
Running a build on the directory that was used to create the source
package will have the patches applied, or running a build on a directory
that was extracted from a source package will have the patches applied.

However, when we run binary-only builds, or autopkgtest directly on the
Debian source repository, no patches are applied, and git needs to run
explicitly.

Make the Quilt run conditional on if a series file exists. If it does
not, the package either has no packages, or the git branch has the
patches already applied (e.g. when building a `gbp pq` branch directly).