Use git rev-parse to retrieve commit id
When using debcraft within a git repository, it is possible that `git
log -n 1 --oneline` can output more than one line, such as when the
configuration value `log.showsignature` is set to show GPG signatures
in the git log.  Because of this, the attempt to assign the `COMMIT_ID`
variable using the command `git -C "$PWD/.git" log -n 1 --oneline | cut
-d ' ' -f 1` results in `COMMIT_ID` being multiple lines.  This causes
problems when it is used as part of the file name used for the image id
file when building the container.

Using the command `git rev-parse --short HEAD` produces the same commit
id as `git log -n 1 --oneline`, and will only contain a single line of
output. Also, it will only output the commit id itself, thus it does not
need to be piped to the `cut` command to get rid of the extra
unnecessary information.