Update build container image to support host arch
This update adds the option `--host-architecture` which sets the
variable `HOST_ARCH`.  This value is used in the container name in order
to distinguish it from containers using other host architectures, as
well as the default container which uses the build architecture as the
host architecture.  The host architecture is also passed to the build
process such that the host architecture is set up as a foreign
architecture, and that the host build dependencies use the host
architecture when resolving and installing build dependencies.