Apply default environment variables on cross build
When performing a cross build, some actions cannot be performed, such as
executing the resulting artifacts utilizing a foreign architecture. This
means that most auto tests fail to run.  Also, some packages may need
special configurations when being cross built.  The `nocheck` option
prevents the check command from running, which is usually used to run
tests.  The `nocheck` and `cross` profiles can be used by packages to
control which dependencies and build processes to use when cross
building and when building without running the check command.

This lines up with how sbuild adds these environment variables when
cross building is dectected, including not overwriting the variables
when they are already set.