Bugfix: Use correct variable name in tests so comparisons actually work
It was discovered that all tests were passing no matter what the string
on the last line was. This was due to the variable having the wrong name
and being empty, so everything passed on comparison to '*'. Fix variable
name and update tests so the last lines are correct and all tests pass.