Ignore quilt exit code 2
Whenever quilt attempts to apply or remove patches, and there are no
patches to apply or remove due to being at the top or bottom of the
stack, quilt give an exit status of 2.  However, `set -e` in a bash
script causes the entire script to stop when it encounters any exit
status other than 0.  Thus, the exit status of 2 needs to be ignored
explicitly.