Make upstream tarball (if exists) available for builds
The `dpkg-source` tool, when run from the source directory
(`/tmp/build/source` inside the container), expects the `.orig.tar.xz`
file to be present in the parent directory (`/tmp/build`). However,
Debcraft was only mounting the source directory and the specific build
directory, but not explicitly copying the upstream tarball into the
build directory. The tarball was downloaded to the main temporary
directory, which was not directly accessible from the container at the
expected relative path.

Fix this by unconditionally attempting to find and copy the
`${PACKAGE}_${PACKAGE_VERSION}.orig.tar.{xz,gz,bz2,lzma}` file into the
build directory if it exists.

Closes: https://github.com/ottok/debcraft/issues/9