When checking for release tags, only consider name space debian|ubuntu/*
This fixes errors with the signature where the upstream/* tag was found
but could not be used as an actual tag to compare to:

    fatal: No tags can describe '0838cfa2240eda2c2eae89bb7935447f0136c6cf'.
    Try --always, or create some tags.

Additionally limit check to current branch (first parents). Debbugs has
native packaging on branch 'master', but it actually never did any
uploads from that branch, but keeps rebasing the branch 'debian'. Thus
tags are found but not on the correct branch and `git describe` can't
actually work.