Amend previous 88360bd to include all bind mounts and all debcraft commands
Mounting of the previous build or previous tagged build are not tested
in the CI and thus this regression was detected only in a local `make
test` run. Also the 'download', 'release' and 'validate' were not properly
tested, but they didn't fail because of how gracefully a shell array
prints as a regular variable.