Switch diffoscope to only compare using the .changes plugin
Stop using diffoscope to compare the build artifacts as such, and just
ask it to compare the *.changes files, using the custom plugin for it has
for the purpose. This seems to give pretty good Debian-specific insight
into the changes.

Also enable the built-in timeout feature in diffoscope. We might be able
to remove the `timeout` wrapper later on if this works well.